refactor:refactor classes to avoid using magic values
