+++ app/src/main/java/com/android/sample/ui/profile/EditProfile.kt
@@ -43,18 +43,20 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.android.sample.model.camera.uploadProfilePicture
+import com.android.sample.model.camera.uriToBitmap
 import com.android.sample.model.profile.ProfileViewModel
 import com.android.sample.model.profile.User
 import com.android.sample.resources.C.Tag.IMAGE_SIZE
 import com.android.sample.resources.C.Tag.LARGE_IMAGE_SIZE
 import com.android.sample.resources.C.Tag.MEDIUM_PADDING
+import com.android.sample.resources.C.Tag.SMALL_PADDING
 import com.android.sample.resources.C.Tag.STANDARD_PADDING
 import com.android.sample.resources.C.Tag.SUBTITLE_FONTSIZE
-import com.android.sample.ui.ImagePicker
-import com.android.sample.ui.ProfileImage
+import com.android.sample.ui.camera.ImagePicker
+import com.android.sample.ui.camera.ProfileImage
 import com.android.sample.ui.navigation.NavigationActions
-import com.android.sample.ui.uploadProfilePicture
-import com.android.sample.ui.uriToBitmap
+
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
