+++ app/src/main/java/com/android/sample/ui/profile/Profile.kt
@@ -53,13 +53,14 @@ import com.android.sample.R
 import com.android.sample.model.activity.ListActivitiesViewModel
 import com.android.sample.model.profile.ProfileViewModel
 import com.android.sample.model.profile.User
+import com.android.sample.resources.C.Tag.IMAGE_SIZE
 import com.android.sample.resources.C.Tag.MEDIUM_PADDING
 import com.android.sample.resources.C.Tag.STANDARD_PADDING
 import com.android.sample.resources.C.Tag.SUBTITLE_FONTSIZE
 import com.android.sample.resources.C.Tag.TEXT_FONTSIZE
 import com.android.sample.resources.C.Tag.TITLE_FONTSIZE
 import com.android.sample.resources.C.Tag.TOP_TITLE_SIZE
-import com.android.sample.ui.ProfileImage
+import com.android.sample.ui.camera.ProfileImage
 import com.android.sample.ui.navigation.BottomNavigationMenu
 import com.android.sample.ui.navigation.LIST_TOP_LEVEL_DESTINATION
 import com.android.sample.ui.navigation.NavigationActions
@@ -181,7 +182,7 @@ fun ProfileContent(
                 // Profile Picture
                 ProfileImage(
                     userId = user.id,
-                    modifier = Modifier.size(100.dp).clip(CircleShape).testTag("profilePicture"))
+                    modifier = Modifier.size(IMAGE_SIZE.dp).clip(CircleShape).testTag("profilePicture"))
 
                 // User Name and Surname
                 Text(
